Generalise bank business logic contract
Extract interface from bank service so that it can express the actions
the bank is able to do.

This allows to write tests in terms of the contract, showing how the
bank is meant to work without coupling them to specific details of a
given layer of the application stack. This will make more sense as soon
as the http api is implemented.

Use gradle test fixtures in order to expose the test to other projects.

It's important to notice, that the each layer of the stack can add extra
tests that are relevant to its context. The tests in terms of the
contract are a baseline and by no means a limitation to the tests that
can be written. In the case of the domain, an additional test shows that
account creation generates its id.
