+++ domain/src/main/kotlin/lmirabal/bank/Bank.kt
@@ -1,0 +1,8 @@
+package lmirabal.bank
+
+import lmirabal.bank.model.BankAccount
+
+interface Bank {
+    fun createAccount(): BankAccount
+    fun listAccounts(): List<BankAccount>
+}
