+++ domain/src/main/kotlin/lmirabal/bank/BankService.kt
@@ -8,11 +8,11 @@ import lmirabal.bank.model.BankAccountId
 class BankService(
     private val accountRepository: BankAccountRepository,
     private val idFactory: () -> BankAccountId = { BankAccountId.random() }
-) {
-    fun createAccount(): BankAccount {
+) : Bank {
+    override fun createAccount(): BankAccount {
         return BankAccount(idFactory(), Amount.ZERO)
             .also { newAccount -> accountRepository.add(newAccount) }
     }
 
-    fun listAccounts(): List<BankAccount> = accountRepository.list()
+    override fun listAccounts(): List<BankAccount> = accountRepository.list()
 }
